RSSI status added (found as part of Negotiate Channel Response from Pump)
Shows user percentage feedback of radio signal to help them diagnose reception issues.

RSSI (Recieved Signal Strength Indicator) is a common name for the signal strength in a wireless network environment. It is a measure of the power level that a RF client device is receiving from an access point, for example.