Bugfix: use SGV timestamp in calculation for next poll.
Previously the PumpStatus' event date was used to calculate the next poll.
Up untill change 045bebb3a2021b523631deec76e8b8d47f7112dc, this date was
set from the SGV date we got from the pump, see
https://github.com/jotomo/600SeriesAndroidUploader/commit/045bebb3a2021b523631deec76e8b8d47f7112dc#diff-20721b0d1f2575ca0cce44c52a932d36L193
After that change, the event date property is the time the record was created.
Hence, now the sgvDate property must be used to acquire the timestamp
of the SGV to calculate the next poll time properly.