Indra - RC2
Improved handling of upload errors.
Nightscout status is read and suggestions offered for changes to settings.
Will work with a NS site that doesn't meet requirements for this version by disabling those elements and informing user.
Removed references to "670G" and only use "Auto Mode" as this mode will be available in as yet unreleased pumps.